% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_finder.R
\name{anime_finder}
\alias{anime_finder}
\title{Application Demo}
\usage{
anime_finder()
}
\value{
Launch the shiny application to give you the ability to search for animes
}
\description{
Function launching the application again
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
