% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_functions.R
\name{cos_similarity}
\alias{cos_similarity}
\title{Compute the cosine similarity between two objects A and B}
\usage{
cos_similarity(A, B)
}
\arguments{
\item{A}{first object}

\item{B}{second object}
}
\value{
Return the cosine similarity value
}
\description{
Compute the cosine similarity between two objects A and B
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
