% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{create_numeric_input}
\alias{create_numeric_input}
\title{Numeric inputs depending on the selection of the user in the selectizeInput}
\usage{
create_numeric_input(
  selectednames,
  selectedcounts,
  min_user = 1,
  max_user = 10,
  placeholder = 5,
  wanted_step = 0.5,
  id
)
}
\arguments{
\item{selectednames}{names recovered through `selectize_names()`}

\item{selectedcounts}{count recovered through `selectize_count()`}

\item{min_user}{minimum value of the numericInputs}

\item{max_user}{maximum value of the numericInputs}

\item{placeholder}{placeholder value in the numericInput}

\item{wanted_step}{step of the value in the numericInput}

\item{id}{id to add when we want to use the functions more than one time in the app}
}
\value{
return a vector containing the numbers the user selected from 1:10
}
\description{
Create numeric inputs depending on the selection of the user in the selectizeInput. You first need to run both `selectize_count()` and `selectize_names()` to run this.
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
