% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_functions.R
\name{item_recommendation}
\alias{item_recommendation}
\title{Use the user item matrix to search for similar users and get the anime they graded the best that we have not watched}
\usage{
item_recommendation(selected_item_name, rating_matrix, n_recommendation, data)
}
\arguments{
\item{n_recommendation}{number of recommendations wanted}

\item{viewed}{user id, set to 999999999 so that there normally is no conflict with other Id's}

\item{dissimularity}{matrix created by the function ...}

\item{all_anim}{table of ratings similar to the one we kept}
}
\value{
Return a table composed of `n_recommendation` that the user have not seen yet
}
\description{
Use the user item matrix to search for similar users and get the anime they graded the best that we have not watched
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
