% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_functions.R
\name{newcommer_recom}
\alias{newcommer_recom}
\title{Filter the anime list depending on simple data}
\usage{
newcommer_recom(anime, age, gender, freetime)
}
\arguments{
\item{anime}{data containing the anime list}

\item{age}{age given by the person}

\item{gender}{gender (type) of interest from the anime data}

\item{freetime}{free time that the person has to use for the anime}
}
\value{
Return a filtered table with the anime depending on the inputs
}
\description{
Filter the anime list depending on simple data
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
