% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{selectize_count}
\alias{selectize_count}
\title{Simply compute the amount of items selected in the selectizeInput}
\usage{
selectize_count(id)
}
\arguments{
\item{id}{inputId of the selectizeInput}
}
\value{
return the number of values selected
}
\description{
Simply compute the amount of items selected in the selectizeInput
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
