% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{selectize_names}
\alias{selectize_names}
\title{Vector of names selected in the selectizeInput}
\usage{
selectize_names(id)
}
\arguments{
\item{id}{inputId of the selectizeInput}
}
\value{
return a vector of the item names selected
}
\description{
Get the vector of names selected in the selectizeInput
}
\examples{

#let's say the user selected those anime in the app
selectize <- c("Naruto", "Death Note", "One Piece")

#we will then recover the following with the function
u <- selectize_names(selectize)

print(u)

}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
