% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_functions.R
\name{user_item_matrix}
\alias{user_item_matrix}
\title{Score per anime per user}
\usage{
user_item_matrix(
  data = anime_with_ratings,
  adding_row = FALSE,
  row_data = NULL
)
}
\arguments{
\item{data}{data having the user ratings per anime}

\item{adding_row}{Variable that is either TRUE if you want the matrix for the user-based recommendation or FALSE if you want the matrix for the item-based recommendation}

\item{row_data}{table created by the function `user_data()`}

\item{anime_with_ratings}{the data by default}
}
\value{
Return a matrix where the x axis include the item_id and the y axis include the user_id. Each line is the score per item per user
}
\description{
Create the user_item matrix from the given data. We use `adding_row = TRUE` for the user based recommendation to be able to add the user selected data to the table
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
