% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_functions.R
\name{item_recommendation}
\alias{item_recommendation}
\title{Similarity between anime}
\usage{
item_recommendation(
  selected_item_name,
  user_item_matrix,
  n_recommendation,
  anime
)
}
\arguments{
\item{selected_item_name}{selection of the user in the Shiny App}

\item{user_item_matrix}{matrix created by the function `user_item_matrix`}

\item{n_recommendation}{number of recommendation wanted by the user}

\item{anime}{is the data used}
}
\value{
Return a table composed of `n_recommendation` that the user have not seen yet
}
\description{
Use the user item matrix to search for similar anime
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
