% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_functions.R
\name{newcommer_recom}
\alias{newcommer_recom}
\title{Filter anime according to preferences}
\usage{
newcommer_recom(anime, age, gender, freetime)
}
\arguments{
\item{anime}{data containing the list of anime with the variables being description for each anime}

\item{age}{age given by the person}

\item{gender}{theme of the anime. For instance if you like action anime, put "action"}

\item{freetime}{free time that the person has to use in order to watch the anime}
}
\value{
Return a filtered table with the anime depending on the inputs
}
\description{
Filter the anime list depending on different variables. These variables must be included in the data set.
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
