% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_functions.R
\name{newcommer_recom}
\alias{newcommer_recom}
\title{Filter anime according to preferences}
\usage{
newcommer_recom(data, age, gender, freetime)
}
\arguments{
\item{data}{data containing the list of anime with the variables being description for each anime}

\item{age}{age given by the person}

\item{gender}{theme of the anime. For instance if you like action anime, put "action"}

\item{freetime}{free time that the person has to use in order to watch the anime}
}
\value{
Return a filtered table with the anime depending on the inputs
}
\description{
Filter the anime list depending on different variables. These variables must be included in the data set.
}
\examples{

#loading the data from the package
anime <- tibble(ProjectG5::anime)

#filtering the table to make propositions to the user through the function
newcommer_recom(anime, 15, "Sport", 30)

}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
