% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{score_recovery}
\alias{score_recovery}
\title{Recover the weights given in the boxes generated by `create_numeric_input()`}
\usage{
score_recovery(selectedcounts, input, id)
}
\arguments{
\item{selectedcounts}{amount of variable selected in the selectize input}

\item{input}{input id from shiny}

\item{id}{id to add when we want to use the functions more than one time in the app}
}
\value{
return a vector containing the numbers the user selected from 1:10
}
\description{
Recover the weights that the user has put in the boxes generated by `create_numeric_input()`
}
\examples{

#let's say the user selected those anime in the app
selectize <- c("Naruto", "Death Note", "One Piece")

#we will use both selectize_names and selectize_count
names <- selectize_names(selectize)
count <- selectize_count(selectize)

#we can then create the inputs
inputs <- create_numeric_input(names, count, id = "example")

#to recover the weight entered by the user we then run
\dontrun{
x <- score_recovery(count, input, id = "example")
}

#note that input are the input from shiny server
#and that this code cannot be ran outside the
#application therefore we do not provide any output here


}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
