% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{selectize_count}
\alias{selectize_count}
\title{Computation of the amount of items selected in the selectizeInput}
\usage{
selectize_count(id)
}
\arguments{
\item{id}{inputId of the selectizeInput}
}
\value{
return the number of values selected

#let's say the user selected those anime in the app
selectize <- c("Naruto", "Death Note", "One Piece")

#we will then recover the following with the function
t <- selectize_count(selectize)

print(t)
}
\description{
Compute the amount of items selected in the selectizeInput
}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
