% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anime_functions.R
\name{user_based_recom}
\alias{user_based_recom}
\title{Similarity between other users to have a list of anime that the user have not seen yet}
\usage{
user_based_recom(
  userid = 999999999,
  user_item_matrix,
  ratings_data,
  n_recommendation = 5,
  threshold = 1,
  nearest_neighbors = 10
)
}
\arguments{
\item{userid}{user id, set to 999999999 so that there is no conflict with other Id's}

\item{user_item_matrix}{matrix created by the function `user_item_matrix`}

\item{ratings_data}{table of ratings similar to the one we kept}

\item{n_recommendation}{number of recommendations wanted}

\item{threshold}{threshold level of number of user that gave a score to the anime}

\item{nearest_neighbors}{number of neighbors taken into account for the computation}
}
\value{
Return a table composed of `n_recommendation` that the user have not seen yet
}
\description{
Use the user item matrix to search for similar users in order to get the anime they graded the best and that we have not watched
}
\examples{

#Simple example to showcase what is happening in the application

#import dataset from the package
anime_with_ratings <-  tibble::tibble(ProjectG5::anime_with_ratings)

#selection of the user example
names_selected <- c("Death Note", "Naruto")
score_entered <- c(10,7)

temp_tibble <- tibble::tibble(Name = names_selected, rating = score_entered)

anime_selected <- dplyr::left_join(anime, temp_tibble, by = c("Name" = "Name"))
anime_selected <- dplyr::filter(anime_selected, Name \%in\% names_selected)
anime_selected <- dplyr::mutate(anime_selected, user_id = 999999999)

#creation of the matrix with what the user selected
item_matrix <- user_item_matrix(anime_with_ratings, adding_row = TRUE, row_data = anime_selected)

user_based_recom(999999999, item_matrix, anime_with_ratings, 5, 1, 10)

}
\author{
Marie Bellier, Massimo Finini, Meri Likoska, Vania Rodrigues Telo Ramos, Xavier Renger
}
